+++ src/modules/linkdrop-p2p/types/get-limits/index.ts
@@ -12,6 +12,6 @@ type TGetLimits = ({
 }: TGetLimitsArgs) => Promise<({
   minTransferAmount: string,
   maxTransferAmount: string
-} | void)>
+})>
 
 export default TGetLimits