+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Linkdrop P2P SDK
 
 ## 2.0.0-beta.0
-- added tokenAddress parameter to getSenderHistory method
+- added "token" parameter to getSenderHistory method
 - bug fix with escrow address for old links
 - operations included to ClaimLink
 - apiKey hidden from properties of ClaimLink