+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Linkdrop P2P SDK
 
+## 2.0.0-beta.0
+- added "token" parameter to getSenderHistory method
+- bug fix with escrow address for old links
+- operations included to ClaimLink
+- apiKey hidden from properties of ClaimLink
+- expiration type changed to 'number'
+
+
+## 2.0.0-alpha.10
+- added getSenderHistory method
+- removed `void` return from public methods
+- added ValidationError
+
 ## 2.0.0-alpha.9
 - update for getLimits method. token argumant is not required for NATIVE tokens
 - bug fix for escrow address detection